+++ xstampp.astpa/src/xstampp/astpa/ui/ATableFilter.java
@@ -15,9 +15,7 @@ package xstampp.astpa.ui;
 
 import org.eclipse.jface.viewers.Viewer;
 
-import messages.Messages;
 import xstampp.astpa.model.ATableModel;
-import xstampp.astpa.model.controlaction.safetyconstraint.ICorrespondingUnsafeControlAction;
 import xstampp.astpa.ui.sds.ModeFilter;
 
 /**
@@ -26,11 +24,9 @@ import xstampp.astpa.ui.sds.ModeFilter;
  * 
  */
 public class ATableFilter extends ModeFilter {
-  private static final String[] FILTERS = new String[] { Messages.All,
-      Messages.CorrespondingSafetyConstraints };
 
   public ATableFilter() {
-    super(FILTERS);
+    super(new String[] {});
   }
 
   @Override
@@ -46,37 +42,9 @@ public class ATableFilter extends ModeFilter {
       if (p.getDescription().toLowerCase().matches(this.searchString)) {
         return true;
       }
-    } else if (element instanceof ICorrespondingUnsafeControlAction) {
-      ICorrespondingUnsafeControlAction cuca = (ICorrespondingUnsafeControlAction) element;
-      return this.checkCSC(cuca);
     }
     return false;
   }
 
-  private boolean checkCSC(ICorrespondingUnsafeControlAction cuca) {
-
-    if ((this.getCSCFilterMode() == 1)) {
-      if (cuca.getCorrespondingSafetyConstraint().getText().toLowerCase()
-          .matches(this.searchString)) {
-        return true;
-      }
-      return false;
-    } else if ((this.getCSCFilterMode() == 2)) {
-      if (cuca.getDescription().toLowerCase().matches(this.searchString)) {
-        return true;
-      }
-      return false;
-    } else if (this.getCSCFilterMode() == 0) {
-      if ((cuca.getCorrespondingSafetyConstraint().getText()
-          .toLowerCase().matches(this.searchString))
-          || (cuca.getDescription().toLowerCase()
-              .matches(this.searchString))) {
-        return true;
-      }
-      return false;
-    } else {
-      return false;
-    }
-  }
 
 }